smart_kitchen_green_server/requirements.txt
Django~=4.2.13
dj-rest-auth
drf-yasg
django-cors-headers
djangorestframework
django-allauth
django-notifications-hq
openmeteo_requests
pandas
setuptools
joblib
requests_cache
retry_requests